+++ src/components/SyntaxHighlighter/index.ts
@@ -1,0 +1,4 @@
+import SyntaxHighlighter from './SyntaxHighlighter';
+
+export { default as SyntaxHighlighterPre } from './CustomPreTag';
+export default SyntaxHighlighter;