+++ src/routes/JobScheduler/views/ScheduledJobForm.tsx
@@ -3,16 +3,24 @@ import Input from '../../../components/Input';
 import { Button, Text, Loader } from '@crate.io/crate-ui-components';
 import Checkbox from '../../../components/Checkbox';
 import SQLEditor from '../../../components/SQLEditor/SQLEditor';
-import { apiPost, apiPut } from '../../../hooks/api';
+import { ApiOutput, apiPost, apiPut } from '../../../hooks/api';
 import { useCallback, useEffect, useState } from 'react';
 import { useGCContext } from '../../../contexts';
 import executeSql, { QueryResults } from '../../../utils/gc/executeSql';
 import SQLResultsTable from '../../../components/SQLResultsTable/SQLResultsTable';
 import { cronParser } from '../../../utils/cron';
 import { Job, JobInput } from '../../../types';
+import cn from '../../../utils/cn';
 
 type ScheduledJobFormAdd = { type: 'add' };
 type ScheduledJobFormEdit = { type: 'edit'; job: Job };
+type JobApiError = {
+  errors: {
+    [key in keyof JobInput]: string[];
+  };
+  message: string;
+  success: boolean;
+};
 
 type ScheduledJobFormProps = {
   backToJobList: () => void;
@@ -30,6 +38,7 @@ export default function ScheduledJobForm(props: ScheduledJobFormProps) {
     handleSubmit,
     setValue,
     watch,
+    setError,
     formState: { errors },
   } = useForm<JobInput>({
     defaultValues:
@@ -45,28 +54,48 @@ export default function ScheduledJobForm(props: ScheduledJobFormProps) {
           },
   });
   const onSubmit: SubmitHandler<JobInput> = async (data: JobInput) => {
-    let result;
+    let result: ApiOutput<JobApiError | Job>;
     setShowLoader(true);
     if (type === 'add') {
       // CREATE
-      result = await apiPost(`${gcUrl}/api/scheduled-jobs/`, data, {
-        credentials: 'include',
-      });
+      result = await apiPost<JobApiError | Job>(
+        `${gcUrl}/api/scheduled-jobs/`,
+        data,
+        {
+          credentials: 'include',
+        },
+        false,
+      );
     } else {
       // UPDATE
       const job = props.job;
-      result = await apiPut(`${gcUrl}/api/scheduled-jobs/${job.id}`, data, {
-        credentials: 'include',
-      });
+      result = await apiPut<JobApiError | Job>(
+        `${gcUrl}/api/scheduled-jobs/${job.id}`,
+        data,
+        {
+          credentials: 'include',
+        },
+        false,
+      );
     }
 
     setShowLoader(false);
     if (result.success) {
       backToJobList();
+    } else if (result.data && 'errors' in result.data) {
+      Object.keys(result.data.errors).forEach(fieldInError => {
+        if (result.data && 'errors' in result.data) {
+          setError(fieldInError as keyof JobInput, {
+            type: 'validate',
+            message: result.data.errors[fieldInError as keyof JobInput].join(' '),
+          });
+        }
+      });
     }
   };
 
   const cron = watch('cron');
+  const sqlValue = watch('sql');
   const cronHumanReadable = cronParser(cron);
 
   const executeQuery = useCallback(() => {
@@ -89,7 +118,10 @@ export default function ScheduledJobForm(props: ScheduledJobFormProps) {
   };
 
   useEffect(() => {
-    register('sql', { required: true, value: type === 'edit' ? props.job.sql : '' });
+    register('sql', {
+      required: 'SQL is a required field.',
+      value: type === 'edit' ? props.job.sql : '',
+    });
   }, [register]);
 
   if (showLoader) {
@@ -103,57 +135,51 @@ export default function ScheduledJobForm(props: ScheduledJobFormProps) {
   return (
     <form
       role="form"
+      id="job-form"
       onSubmit={handleSubmit(onSubmit)}
       className="flex flex-col gap-4"
     >
-      <div className="grid grid-cols-12 gap-4">
-        <div className="col-span-6 flex flex-col gap-2">
+      <div className="flex flex-col md:grid md:grid-cols-12 gap-2 md:gap-4">
+        <div className="col-span-5 flex flex-col gap-2">
           <Input
-            {...register('name', { required: true })}
+            {...register('name', { required: 'Job Name is a required field.' })}
             required
             id="name"
             label={'Job Name'}
-            error={
-              errors.name && (
-                <Text className="text-red-500">Job Name is a required field.</Text>
-              )
-            }
+            error={errors.name && errors.name.message}
           />
         </div>
 
         <div className="col-span-5 flex flex-col gap-2">
           <Input
             {...register('cron', {
-              required: true,
+              required: 'Schedule is a required field.',
               validate: (value: string) => {
-                return cronParser(value) !== null;
+                return cronParser(value) !== null ? true : 'Invalid CRON schedule.';
               },
             })}
             id="cron"
             required
             label={'Schedule'}
             placeholder="* * * * *"
-            error={
-              errors.cron && errors.cron.type === 'required' ? (
-                <Text className="text-red-500">Schedule is a required field.</Text>
-              ) : errors.cron && errors.cron.type === 'validate' ? (
-                <></>
-              ) : undefined
-            }
+            error={errors.cron && errors.cron.message}
           />
 
-          {cronHumanReadable === null &&
-          !(errors.cron && errors.cron.type === 'required') ? (
-            <Text>Invalid CRON schedule.</Text>
-          ) : cronHumanReadable !== null ? (
-            <Text>
-              This job will run{' '}
-              <span className="font-bold">{cronHumanReadable.toLowerCase()}</span>.
-            </Text>
-          ) : null}
+          <Text
+            className={cn({
+              invisible: cronHumanReadable === null,
+              hidden: errors.cron,
+            })}
+          >
+            This job will run{' '}
+            <span className="font-bold">
+              {cronHumanReadable && cronHumanReadable.toLowerCase()}
+            </span>
+            .
+          </Text>
         </div>
 
-        <div className="col-span-1 flex flex-col gap-2 pt-8">
+        <div className="col-span-2 flex flex-col gap-2 md:pt-8">
           <Checkbox {...register('enabled')} label={'Active'} id="job_active" />
         </div>
       </div>
@@ -162,31 +188,24 @@ export default function ScheduledJobForm(props: ScheduledJobFormProps) {
         <SQLEditor
           results={queryResults}
           localStorageKey="sql-job-editor"
-          runButtonLabel="Test Queries"
-          value={type === 'edit' ? props.job.sql : ''}
+          value={sqlValue || ''}
           onChange={query => {
             setValue('sql', query, { shouldValidate: true });
           }}
           onExecute={() => {
             executeQuery();
           }}
-          error={
-            errors.cron && (
-              <Text className="text-red-500">SQL is a required field.</Text>
-            )
-          }
+          error={errors.sql && errors.sql.message}
         />
       </div>
 
-      <div className="flex justify-between">
+      <div className="w-full flex flex-col md:flex-row justify-end gap-2">
         <Button kind={Button.kinds.SECONDARY} onClick={backToJobList}>
           Cancel
         </Button>
-        <div className="flex gap-2">
-          <Button kind={Button.kinds.PRIMARY} type={Button.types.SUBMIT}>
-            Save
-          </Button>
-        </div>
+        <Button kind={Button.kinds.PRIMARY} type={Button.types.SUBMIT}>
+          Save
+        </Button>
       </div>
 
       <div className="mt-4">{renderResults()}</div>