+++ src/types/job.ts
@@ -8,13 +8,36 @@ export type Job = {
   last_execution?: JobLog;
 };
 
+export type TJobLogStatementError = {
+  error: string;
+  sql: string;
+};
+export type TJobLogStatementSuccess = {
+  duration: number;
+  sql: string;
+};
+
+export type TJobLogStatementsError = {
+  [key: string]: TJobLogStatementError;
+};
+export type TJobLogStatementsSuccess = {
+  [key: string]: TJobLogStatementSuccess;
+};
+
+export type ErrorJobLog = {
+  error: string;
+  statements: TJobLogStatementsError;
+};
+export type SuccessJobLog = {
+  error: null;
+  statements: TJobLogStatementsSuccess;
+};
+
 export type JobLog = {
   job_id: string;
   start: string;
   end: string;
-  error: string | null;
-  statements: unknown;
-};
+} & (ErrorJobLog | SuccessJobLog);
 
 export type JobInput = {
   name: string;